LPS-107120 Moves code to LayoutDataConverter class to reuse it
