--- modules/apps/layout/layout-taglib/src/main/java/com/liferay/layout/taglib/internal/util/RenderFragmentLayoutTagUtil.java
@@ -1,36 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.layout.taglib.internal.util;
-
-import com.liferay.portal.kernel.json.JSONObject;
-
-/**
- * @author Eudaldo Alonso
- */
-public class RenderFragmentLayoutTagUtil {
-
-	public static final int LATEST_VERSION = 1;
-
-	public static boolean isReactEditor(JSONObject dataJSONObject) {
-		int version = dataJSONObject.getInt("version");
-
-		if (version == LATEST_VERSION) {
-			return true;
-		}
-
-		return false;
-	}
-
-}
